+++ indico/modules/rb/operations/bookings.py
@@ -395,7 +395,7 @@ def should_split_booking(booking, new_data):
     return is_ongoing_booking and (times_changed or repetition_changed or weekdays_changed)
 
 
-def split_booking(booking, new_booking_data):
+def split_booking(booking, new_booking_data, request_data):
     is_ongoing_booking = booking.start_dt.date() < date.today() < booking.end_dt.date()
     if not is_ongoing_booking:
         return
@@ -425,7 +425,7 @@ def split_booking(booking, new_booking_data):
             'recurrence_weekdays': booking.recurrence_weekdays,
         }
 
-        booking.modify(old_booking_data, session.user)
+        booking.modify(old_booking_data, session.user, request_data=request_data)
 
     for occurrence_to_cancel in occurrences_to_cancel:
         occurrence_to_cancel.cancel(session.user, silent=True)